refactor wikipedia feature
