+++ src/features/map/MapContent.jsx
@@ -3,7 +3,7 @@ import { useMap } from "react-leaflet";
 import { useSelector } from "react-redux";
 import Wikipedia from "./layers/Wikipedia";
 import { selectCourseLine, selectCurrentMap, selectIsFollowing, selectVisualizeSearchRadius } from "./mapSlice";
-import { selectIsEnabled } from '../wikipedia/wikipediaSlice';
+import { selectIsEnabled } from '../wikipedia/wikipediaSelectors';
 import { MainLayer } from "./layers/MainLayer";
 import { selectSimdata } from "../simdata/simdataSlice";
 import CourseLine from "./layers/CourseLine";
