+++ src/features/map/layers/SearchRadiusCircle.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Circle } from 'react-leaflet';
 import { useSelector } from 'react-redux';
 import { geolibToArrayPoint } from '../../../utils/geo';
-import { selectSearchCenterPoint, selectSearchRadius } from '../../wikipedia/wikipediaSlice';
+import { selectSearchCenterPoint, selectSearchRadius } from '../../wikipedia/wikipediaSelectors';
 
 export default function SearchRadiusCircle() {
   const searchRadius = useSelector(selectSearchRadius);
