+++ src/features/wikipedia/wikipediaSlice.js
@@ -1,12 +1,8 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import repository from './repository';
 import wikipediaEditions from './wikipediaEditions';
-import { createSelector } from 'reselect';
-import { computeDestinationPoint, getDistance, getRhumbLineBearing } from 'geolib';
+import { getDistance, getRhumbLineBearing } from 'geolib';
 import { angleDiff, arrayToGeolibPoint, wikipediaPointToGeolibPoint } from '../../utils/geo';
-import { selectSimdata } from '../simdata/simdataSlice';
-import { selectAutoPlay, selectIsPlaying, toggleIsPlaying } from '../tts/ttsSlice';
 
 export const wikipediaSlice = createSlice({
   name: 'wikipedia',
@@ -156,140 +152,5 @@ export const {
   updateCalculatedData,
   updateLastSearch
 } = wikipediaSlice.actions;
-
-function pageSort(a, b) {
-  let res = a?.closestPoint && b?.closestPoint ? a.closestPoint.distance - b.closestPoint.distance : 0;
-  if (res === 0) {
-    res = a.page.title.localeCompare(b.page.title);
-  }
-  if (res === 0) {
-    res = a.page.pageid - b.page.pageid;
-  }
-  return res;
-}
-
-export const getPages = (lat, lng, radius) => async (dispatch, getState) => {
-  dispatch(updateLastSearch({
-    searchTime: new Date().getTime(),
-    searchPosition: [lat, lng],
-    searchRadius: radius
-  }));
-  let state = getState();
-  const edition = selectEdition(state);
-  const data = await repository.getPagesByGeoLocation(edition, lat, lng, radius);
-  if (data) {
-    dispatch(receivePages({
-      data, 
-      searchPosition: [lat, lng],
-      searchRadius: radius,
-      searchTime: new Date().getTime()
-    }));
-    state = getState();
-    const {
-      position,
-      heading
-    } = selectSimdata(state);
-    dispatch(updateCalculatedData({ position, heading, currentTime: new Date().getTime() }));
-    const playingPage = selectPlayingPage(state);
-    if (!playingPage) {
-      dispatch(playNext());
-    }
-  }
-};
-
-export const clearPagesOutOfRange = () => (dispatch, getState) => {
-  const maxPagesInState = 100;
-  const maxDistanceBehindPlayerBeforePageIsRemoved = 40000;
-  
-  const state = getState();
-  const pages = selectPagesWithDistances(state);
-  const pagesToRemove = pages.filter((p, i) => {
-    if (i >= maxPagesInState && state.wikipedia.playingPageid !== p.page?.pageid) {
-      return true;
-    }
-    return !p?.closestPoint?.isInFront && 
-            p?.closestPoint?.distance > maxDistanceBehindPlayerBeforePageIsRemoved;
-  });
-  dispatch(removePages({ pageids: pagesToRemove.map(p => p.page?.pageid) }));
-};
-
-export const playNext = () => (dispatch, getState) => {
-  const state = getState();
-  const autoPlay = selectAutoPlay(state);
-  const isPlaying = selectIsPlaying(state);
-  if (!autoPlay && isPlaying) {
-    dispatch(toggleIsPlaying());
-  }
-  const allPages = selectPagesWithDistances(state).filter(p => p.page.pageid !== state.wikipedia.playingPageid);
-  dispatch(wikipediaSlice.actions.playNext({ nextPageid: allPages[0]?.page?.pageid }));
-};
-
-export const selectIsEnabled = (state) => state.wikipedia.isEnabled;
-
-export const selectEdition = (state) => state.wikipedia.edition;
-
-export const selectAvailableEditions = (state) => state.wikipedia.availableEditions;
-
-export const selectPages = (state) => state.wikipedia.pages;
-
-export const selectSearchRadius = (state) => state.wikipedia.searchRadius;
-
-export const selectLastSearchPosition = (state) => state.wikipedia.lastSearchPosition;
-
-export const selectLastSearchRadius = (state) => state.wikipedia.lastSearchRadius;
-
-export const selectLastSearchTime = (state) => state.wikipedia.lastSearchTime;
-
-export const selectPagesWithDistances = createSelector(
-  (state) => ({
-    pages: state.wikipedia?.pages,
-    calculatedData: state.wikipedia?.calculatedData,
-    playingPageid: state.wikipedia?.playingPageid
-  }),
-  ({ pages, playingPageid, calculatedData }) => {
-    const pagesWithClosestPoints = pages?.map(p => {
-      const { closestPoint } = calculatedData[p.pageid] ?? {};
-
-      return {
-        page: p,
-        closestPoint,
-        isReading: playingPageid === p.pageid
-      };
-    }).sort(pageSort);
-    
-    return [
-      ...pagesWithClosestPoints.filter(p => p.closestPoint?.isInFront),
-      ...pagesWithClosestPoints.filter(p => !p.closestPoint?.isInFront),
-    ];
-  }
-);
-
-export const selectPlayingPage = createSelector(
-  (state) =>  ({
-    pages: selectPagesWithDistances(state),
-    playingPageid: state.wikipedia?.playingPageid
-  }),
-  ({ pages, playingPageid }) => {
-    return pages.find(p => p.page.pageid === playingPageid);
-  }
-);
-
-export const selectSearchCenterPoint = createSelector(
-  (state) => ({
-    position: state.simdata?.position,
-    heading: state.simdata?.heading,
-    speed: state.simdata?.airspeed,
-    searchRadius: state.wikipedia?.searchRadius
-  }),
-  ({ position, heading, searchRadius, speed }) =>
-    position ? (
-      speed > 50 ?
-      computeDestinationPoint(
-        arrayToGeolibPoint(position),
-        searchRadius, 
-        heading
-      ) : arrayToGeolibPoint(position)
-    ) : undefined
-);
 
 export default wikipediaSlice.reducer;
